brain-api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
asyncpg>=0.30.0
redis[hiredis]>=5.0.0
httpx>=0.28.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
